+++ flexmeasures/__init__.py
@@ -1,1 +1,12 @@
-__version__ = "0.3.2"
+from importlib_metadata import version, PackageNotFoundError
+
+
+__version__ = "Unknown"
+
+# This uses importlib.metadata behaviour added in Python 3.8
+# and relies on setuptools_scm.
+try:
+    __version__ = version("flexmeasures")
+except PackageNotFoundError:
+    # package is not installed
+    pass